Don't store assignments when the segmenter returns nil?
When a segmented returns nil, it means "unqualified for now", which means Verdict will send the subject to the segmenter again when it returns. This allows you to defer assignments, which is nice for rollouts scenarios.

Contrast this with the qualifier block, which permanently disqualifies a subject from the experiment (if store_unqualified is not set to false).